Use Packer/Unpacker for codec spi
